feat: added subExample of actionsheet component
