+++ package.json
@@ -10,7 +10,6 @@
     "test": "jest --watchAll",
     "extract-code": "node scripts/extractCode.js",
     "extract-examples": "node scripts/extractExamples.js",
-    "extract-actionsheet-examples": "node scripts/extractActionsheetExamples.js",
     "add-imports": "node scripts/addImports.js",
     "extract-docs-examples": "node scripts/extractCode.js; node scripts/extractExamples.js; node scripts/addImports.js"
   },
